refactor: remove StartupQueue and move logic to DeviceModeDestinationPlugin
